added enc_password
